+++ annotations/nocedit.exe/pseudocode/src/core/charactr.cpp/CCharacter_processDamage_FUN_0042c3c0.keep.cpp
@@ -83,7 +83,7 @@ LAB_0042c429:
   if ((this_ptr->health_bar_mode == 1) && ((CHero *)this_ptr != g_HeroActors[g_LocalHeroIndex])) {
     core_game_cpp_CGame_setStatusDisplay_FUN_004e0bf0
               (g_CGamePtr,this_ptr->descriptive_name,
-               (int)(this_ptr->hit_points / this_ptr->max_hit_points),5.0);
+               this_ptr->hit_points / this_ptr->max_hit_points,5.0);
   }
   if (0.0 < this_ptr->hit_points) {
     return;
