+++ annotations/nocedit.exe/pseudocode/src/core/vampboss.cpp/CVampireBoss_process_FUN_005e5970.keep.cpp
@@ -585,7 +585,7 @@ LAB_005e5f5f:
   if ((pSVar19->state_index != 2) && (g_CGamePtr->letterbox_mode == 0)) {
     core_game_cpp_CGame_setStatusDisplay_FUN_004e0bf0
               (g_CGamePtr,(this_ptr->base).base.descriptive_name,
-               (int)((this_ptr->base).base.hit_points / (this_ptr->base).base.max_hit_points),0.5);
+               (this_ptr->base).base.hit_points / (this_ptr->base).base.max_hit_points,0.5);
     return;
   }
   return;
