+++ annotations/nocturne.exe/pseudocode/src/core/charactr.cpp/CCharacter_process_FUN_004259f0.cpp
@@ -37,7 +37,7 @@ int __cdecl core_charactr_cpp_CCharacter_process_FUN_004259f0(CCharacter *this_p
   if (this_ptr->health_bar_mode == 2) {
     core_game_cpp_CGame_setStatusDisplay_FUN_004a3ab0
               (g_CGame_PTR_005b9354,this_ptr->descriptive_name,
-               (int)(this_ptr->hit_points / this_ptr->max_hit_points),2.0);
+               this_ptr->hit_points / this_ptr->max_hit_points,2.0);
   }
   if (this_ptr->render_active == 0) {
     if (this_ptr->closest_distance_threshold == 9999.0f) {
